FrameWork Created and Test case 1 and 2 finished
