--- src/test/java/FeatureFiles/_02_Ability_to_Login_User.feature
@@ -1,10 +1,0 @@
-Feature:Ability to Login User
-
-  Background:
-    Given  Navigate to Site
-    When Click on Sing in Button
-  Scenario:
-    Then Enter E-mail adress And Password
-    And Click on Sing in
-
-
