+++ src/test/java/FeatureFilesSerkanSengul/_02_Ability_to_Login_UserSerkanS.feature
@@ -1,0 +1,8 @@
+Feature:Ability to Login User
+
+  Background:
+    Given  Navigate to Site
+    When Click on Sing in Button
+  Scenario:
+    Then Enter E-mail adress And Password
+    And Click on Sing in
