+++ src/test/java/SerkanSengul/ParentSerkanSengul.java
@@ -1,0 +1,59 @@
+package SerkanSengul;
+
+
+import Utilities.GWD;
+import org.openqa.selenium.JavascriptExecutor;
+import org.openqa.selenium.Keys;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.interactions.Actions;
+import org.openqa.selenium.support.ui.ExpectedConditions;
+import org.openqa.selenium.support.ui.WebDriverWait;
+import org.testng.Assert;
+
+import java.time.Duration;
+
+public class ParentSerkanSengul {
+    public WebDriverWait wait = new WebDriverWait(GWD.getDriver(), Duration.ofSeconds(20));
+    Actions actions = new Actions(GWD.getDriver());
+
+    public void Myclick(WebElement element) {
+        wait.until(ExpectedConditions.elementToBeClickable(element));
+        scrollToElement(element);
+        element.click();
+    }
+
+    public void MysendKeys(WebElement element, String yazi) {
+        wait.until(ExpectedConditions.visibilityOf(element));
+        scrollToElement(element);
+        element.clear();
+        element.sendKeys(yazi);
+    }
+
+    public void scrollToElement(WebElement element) {
+        JavascriptExecutor js = (JavascriptExecutor) GWD.getDriver();
+        js.executeScript("arguments[0].scrollIntoView();", element);
+    }
+
+    public void VerifycontainsText(WebElement element, String value) {
+        wait.until(ExpectedConditions.textToBePresentInElement(element, value));
+        Assert.assertTrue(element.getText().toLowerCase().contains(value.toLowerCase()));
+        actions.sendKeys(Keys.ESCAPE).perform();
+    }
+
+
+    public void SuccessfullyDeletedMesage(WebElement element, String value) {
+        wait.until(ExpectedConditions.textToBePresentInElement(element, value));
+        Assert.assertTrue(element.getText().toLowerCase().contains(value.toLowerCase()));
+        actions.sendKeys(Keys.ESCAPE).perform();
+
+    }
+
+
+
+    }
+
+
+
+
+
+
